*Compile portlibs with devkitPPC r24 and libogc 1.8.8 *Move the stack up 8 bytes and align 32 after bss also, so that .end is always aligned. *Fix finding of the main application in channels, there is not need for a second scan of the app files and checking for a wad header with memcmp if you get the info from the TMD correctly.  This eliminates the skip_bootcontent method and all nand games boot the same way. *Add a workaround for custom channels (forwarders) so that they use WII_LaunchTitle instead of the custom appldr. *Fix booting of injected vc. *Align everything going through ioctls and ioctlv for nand emu, to improve cache flushing. *Improve the nand emu mount/dismount and simplify it. *Rename everything WiiTDB to GameTDB, and change to the new URL's.  This includes some fixes to the cover downloader and speeds it up ALOT.  Instead of 12 url requests for each missing cover, now it will be 2-3 at most.  There are also no longer seperate url's for channels and games.  Also renamed any wiitdb references in languages and themes to the gametdb counterpart, using the same case as it were before. *Some cleanup of wbfs initialization, removing some unneeded or unused parameters and variables. *Added sdhc.h and sdhc.c from CFG *Added support for playing games from SD with the same file system support as USB. *Added sorting by console type based on the id of the game.  Use it by sorting By Game ID and using b+up/down to cycle through the systems. --Some new translations for this: custom=Custom commodore=Commodore 64 neogeo=Neo-Geo nes=Nintendo snes=Super Nintendo mastersystem=Sega Master System genesis=Sega Genesis nintendo64=Nintendo64 turbografx16=TurboGrafx-16 turbografxcd=TurboGrafx-CD wiiware=WiiWare wiichannels=Offical Wii Channels wii=Wii homebrew=Homebrew unknown=Unknown
*Lower mem2 priority size to send more shit to MEM2_alloc
*Raise mem1 upper limit for the apploader 8 bytes.
*Make cheat downloader work for WiiWare and VC
*Delete menu 7, forgot last commit
*Attempt at fixing case colors, still needs work, also titles may need some work since they seem to not be read form xml or custom_titles.ini sometimes.
*Clarify the setting up of low memory in disc.c
*Remove the ios parameter from the apploader, Disc_BootPartition, and Disc_WiiBoot.
*Make frag truly work for WBFS partitions.
*Clean up some includes throughout.
*Make all data pointers and parameters for channel launching use u8 * instead of u32 * to get rid of mass casting in pointer math.
*A shitload more im sure, need to commit to sync to official SVN.







git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@417 43c7226f-5572-e80b-0314-b0280708705f

+++ source/channel/channels.h
@@ -54,9 +54,9 @@ class Channels
 
 		void Init(u32 channelType, string lang, bool reload = false);
 
-		u32 * Load(u64 title, char* id);
+		u8 * Load(u64 title, char* id);
 		u8 GetRequestedIOS(u64 title);
-		bool Launch(u32 *data, u64 chantitle, u8 vidMode, bool vipatch, bool countryString, u8 patchVidMode);
+		bool Launch(u8 *data, u64 chantitle, u8 vidMode, bool vipatch, bool countryString, u8 patchVidMode);
 		
 		int Count();
 		wchar_t *GetName(int index);
@@ -74,7 +74,7 @@ class Channels
 		
 		static int GetLanguage(const char *lang);
 		u64* GetChannelList(u32* count);
-		static bool GetAppNameFromTmd(u64 title, char* app, bool dol = false, u16* bootcontent = NULL);
+		static bool GetAppNameFromTmd(u64 title, char* app, bool dol = false, u32* bootcontent = NULL);
 		static bool GetChannelNameFromApp(u64 title, wchar_t* name, int language);
 
 		void Search(u32 channelType, string lang);
